Smooth. Fix an issue with prerendering when there was no user
Nothing bad happend UX wise, but we did a get&post on an event when there was no user present.
Now we check if there's a user on the session, and then always retrieve the event. Cause the
event might be needed outside the admin page, like the user. For showing the preview button
for example, or showing the "sign up" message with a "go to event" text.